#2292 Unit tests for comparer
